regular_stacks/requirements.txt
jadecobra
aws-cdk-lib
constructs
sniffer
macfsevents
aws_cdk.aws_apigatewayv2_alpha
aws_cdk.aws_apigatewayv2_integrations_alpha
well_architected_constructs
well_architected_stacks